requirements.txt
aniso8601==9.0.1
attrs==22.2.0
autopep8==2.0.0
beautifulsoup4==4.12.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coverage==6.5.0
exceptiongroup==1.1.0
flasgger==0.9.5
Flask==2.2.3
Flask-Pydantic==0.11.0
Flask-RESTful==0.3.9
Flask-WTF==1.1.1
future==0.18.3
html5lib==1.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
lxml==5.2.2
MarkupSafe==2.1.2
mistune==2.0.5
packaging==23.0
peewee==3.16.2
pefile==2021.9.3
pluggy==1.0.0
pycodestyle==2.9.1
pydantic==1.10.8
pydentic==0.0.1.dev3
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.2.1
pytest-cov==4.0.0
pytest-mock==3.10.0
python-dotenv==1.0.0
python-stdnum==1.18
pytz==2023.3
PyYAML==6.0
requests==2.28.1
six==1.16.0
soupsieve==2.4
tomli==2.0.1
typing_extensions==4.6.2
urllib3==1.26.12
webencodings==0.5.1
Werkzeug==2.2.3
WTForms==3.0.1
